requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
supabase==2.13.0
pytrends==4.9.2
pandas==2.2.3
numpy==2.2.1
scipy==1.15.0
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
google-analytics-data==0.18.16
google-auth==2.37.0
jinja2==3.1.5
weasyprint==63.1
apscheduler==3.11.0
